Replace ZoneRegistry by placing and locating ZoneValidators via Cluster Sharding
The peristence ID used for ZoneValidators is now changed to be the string of the
ZoneId rather that just its UUID, i.e. ZoneId(abc...) as opposed to just abc...
- this is more future proof  - in case other entities are added that would have
collided if named only with a UUID-like string.

Also there was an issue with ZoneValidator termination before whereby:

a) ClientConnection did not receive Terminated(_) because Terminated is only
   sent when an actor has stopped permanently (i.e. it will not be restarted),
   but we were watching it in case of crashes (creating a crash with
   sys.error("test") in the ChangeZoneNameCommand handler showed this to be the
   case).
b) Even if it had received Terminated(_), the client would never have recieved
   the ZoneTerminatedNotification because ClientConnection was not converting it
   from a JsonRpcNotificationMessage to a JSON string, and inspection of Play!'s
   websocket actor confirms they'd have just been dropped. So the client would
   remain connected, oblivious to the fact it was no longer a member of the
   restarted ZoneValidator and that all commands it sent were being silently
   dropped by ClientConnection (as foreach on an empty Option does nothing).

This is fixed by using the PubSub extension to enable ClientConnections to
subscribe to ZoneId-named-topics in which ZoneValidators broadcast lifecycle
events (preStart and postStop).

The update to liquidity-common 0.29.0 just enables the use of a default None
parameter when creating error responses.