Add ZoneValidatorSpec
(Requires a running Cassandra Docker container at 172.17.0.2).

cassandra-journal.contact-points should actually never have been with the akka {
} config block, but there's no point rewriting history to change that. I only
noticed because for some reason the setting of the contact point to "cassandra"
in application.conf works even when it's in the akka {} block, but the override
in test.conf does not work that way. This seems to be related to
https://github.com/krasserm/akka-persistence-cassandra/pull/116, so it may have
been that the proper location changed in v0.6.