Add upgrade smoke integration test
The zone ID (""4cdcdb95-5647-4d46-a2f9-a68e9294d00a"") hardcoded in this new
test is the ID of a production zone. The test serves to validate that
upgrades of components in the persistence path (i.e. the
akka-persistence-cassandra plugin, and Cassandra itself) preserve existing
production data.

The three tests pass the first time they are run against a local Docker
container that has had a copy of the production data imported. They fail on
subsequent runs (and this is expected) because on subsequent runs the restarted
zone validators send ZoneRestarted messages to their clients -- and as the
TestKit's actors path is stable, it ends up receiving them, which interferes
with the test expectations. So for now the production data has to be freshly
imported before each test suite run.