Add ApplicationSpec to test WebSocket API
There's a bit of refactoring that could be done, the the PEM string used as the
certificate header differs slightly from what NGINX provides (it replaces all
but the first newline with tabs), and it could be more extensive
(as could ZoneValidatorSpec)....

Nevertheless this is a big step forward, as it enables quick regression testing
of Play etc. updates, so is worth committing as is. It can be improved later.