Convert ServerConnectionSpec from integration test to unit test
It used to run against an instance of LiquidityServer, which entailed
starting a Cassandra instance and made the test unnecessarily
heavyweight.

Now it runs against lightweight mocks of _just_ the required
functionality. This also enables assertions on the "other side", but the
main benefit is the decreased execution time.

The other changes were all required to support this. Specifically the
changes to ServerTrust and ClientKey were prompted because their
previously impure states meant that ServerConnectionSpec failed when run
as part of the entire client module test suite because they remained
initialised after having been used in their own tests (so the
re-initialisation was not having the effect it should have had). Hence
the elimination of global state from them (inspection of the Android
client showed that they are already no longer ever accessed directly;
the ServerConnection instance encapsulates all access, and so the
singleton functionality of ServerTrust and ClientKey themselves has been
redundant for some time). The change to ServerConnection to allow
directly creating new instances combines with the changes to ServerTrust
and ClientKey to allow the tests to be avoid creating or reusing _any_
global state (again, inspection of the Android client showed that the
client is now already taking responsibility for singleton creation and
reuse, so the functionality of ServerConnection itself has been
redundant for some time).