Define and migrate to Protobuf serialization for persistence
Read and write EventAdapters are used in order to isolate the majority
of the codebase from this change.

The rename of persistence.Event to persistence.ZoneEvent was not
_required_ to achieve this, and is instead done only to improve naming
(doing it as part of this change saves repeating the test effort).

The migration (fully tested against production data in a local throwaway
environment) is included in the main (and only) server image; it is to
be run via `docker-compose up journal-migration` (after having first
stopped the other containers). Once it has been run, new containers can
be launched.  The migration is pure in the sense that it doesn't alter
the original keyspace at all; instead, it writes out the converted
events to a new journal in a new keyspace.

It's worth noting that his change reduces the size of the journal data
to about 2/3 of what it was (although that was not the primary
motivation for the change).