Convert ZoneValidator actor protocol to Protobuf
The ZoneValidator protocol is now:
1. Independent of the WsProtocol.
2. Still represented by the same types that they always were, as far as speakers
   of the protocol are concerned.
3. Serialized to Protobuf messages.

(3) was the goal of this change, while (2) was a constraint on the solution that
I've imposed in order to retain meaningful types throughout most of the codebase
without resorting to using ScalaPB's TypeMappers (which can only do so much).

The introduction of several new `case object Empty ...` types on the application
side is in order to enable inductive derivation of ProtoConverter instances for
the wrapper coproducts that ScalaPB generates for `oneof` message types.

The movement of ZoneValidatorMessageSerializer fromm the actor-protocol module
to the server module was done in order ensure that its compilation happens
_after_ the compilation of both the generated Protobuf protocol types and the
original protocol types (both of which are compiled in the actor-protocol
module). This was necessary because ScalaPB generates the _instances_ of the
wrapper coproducts that it creats for `oneof` message types _within_ the
companion object that it creates for the trait of which they are children.
This property of the generated code interacted badly with the way that
scalac's knownDirectSubclasses macro method works, causing errors of the form
`error: knownDirectSubclasses of Foo observed before subclass Bar registered`.

I found that commenting out the three lines in ZoneValidatorMessageSerializer
that trigger the derivation of ProtoConverter instances that depend on coproduct
ProtoConverter instances being derived was sufficient for compilation to
succeed. I then found that a second compilation attempt with them _un_commented
would somewhat suprisingly also succeed. This finding coupled with the notes on
https://github.com/scala/scala/pull/5284 confirmed to me that the issue was
only happening because of the above property of the generated ScalaPB code
(ProtoConverterSpec includes test cases for coproducts which work fine, so I
knew that other derivations were working - although somewhat surprisingly it
also works if the test coproduct instances are moved to be children of
companions of their parent traits...).

Hence the workaround of simply moving the serializer instances into the server
module (which is the only place they're used anyway*). This ensures that
derivation happens _after_ the necessary compiler phases have completed on the
coproducts in question.

*ZoneEventSerializer was in a separate module because analytics used to run as a
separate deployable which built against the persistence module (but not the
server module).

Other notes:

1. There are two TODOs re. unused import warnings in the ZoneEvent and
   ZoneValidatorMessage serializers. I've left these for later resolution
   because: (a) it's taken a fair amount of time already to get this change to
   where it is, (b) the size of the diff is getting large, and (c) they aren't
   issues with the code anyway - they really seem to be an IntelliJ issue, so
   the TODOs are really about finding a fix or workaround for that.
2. ClientConnectionActor (and its Spec) are using ProtoConverter instances to
   convert WsProtocol ZoneCommands to ZoneValidatorMessage ZoneCommands. As the
   numerous comments where this occurs explain, the terms asScala and asProto
   are not appropriate here. Indeed, that's not the only way this is not ideal.
   The more preferable solution would have been to _move_ the ZoneCommands from
   WsProtocol, rather than _copy_ them. That's actually what I first attempted
   to do. However, the issue I ran into was in adapting the scala-json-rpc
   message format constructors to support wrapping an external message in a way
   that would preserve the original JSON-RPC wire format. There would have been
   a solution, but using the recently developed ProtoConverter functionality
   ended up being a much faster solution to implement (and is also less
   verbose). Since the JSON-RPC protocol is now living on borrowed time and
   should be deprecated and then removed in the near future now that all the
   prerequisite steps are complete (this being the last of them), this solution
   is reasonable enough (it will disappear when the JSON-RPC protocol is
   removed). This leads nicely on to the final point:
3. play-json removal: as part of the deprecation and later removal of the
   JSON-RPC protocol, the number of modules which include play-json as a
   dependency should decrease and eventually reach zero. For now it remains as
   a dependency of most things because I've deliberately left the application
   side ADTs with JsObject as the metadata type in order to minimise the volume
   of individual Protobuf-conversion-related commits.