Update to okhttp 3.8.0
okhttp-ws 3.4.2 was the last release where websocket support was a
separate module, hence the existence of later releases - as part of
okhttp itself - not being picked up until now.

I only realised now because in enabling strict dependency conflict
resolution I found that okhttp-ws was depending on an older version of
okio than the model module, and so looked into the state of okhttp.