Move ZoneId from commands to envelope, remove WsProtocol zone commands
The ZoneValidator actor-protocol is now considered part of the client
API. This removes the duplicity of definitions and the need for mapping
between them that existed before.

LegacyClientConnectionActor is updated to no longer use ProtoBindings,
as a result of the legacy WsProtocol zone commands no longer matching
the reworked ZoneValidator zone commands.

WsProtocol is removed in favour of using the WsProtocol proto types
directly, because the cost of ProtoBindings isn't worth the benefit in
this specific case.