Disallow Java serialization
Even with `extends NoSerializationVerificationNeeded` marking all
local-only messages, there were still false negatives in tests for e.g.
scala.runtime.BoxedUnit, due to the use of serialize-messages = on, for
which there is no obvious supression mechanism.

Given that the multi-jvm test triggers serialization of those messages
which really require verification anyway, the conclusion reached and
effected here is that it's preferable to set allow-java-serialization =
off at the expense of being able to set serialize-messages = on and
serialize-creaters = on*. Given there is no longer any need for it, the
use of the NoSerializationVerificationNeeded marker is now removed.

*this wasn't causing false negatives, but I also concluded that there is
no real point in setting it, given that *every* actor in the project has
local deploy scope, and so they were specifically marked as such (those
newly redundant deploy scope markers are also removed in this change,
and the safer - in terms of preserving evaluation strategies - props
pattern is restored throughout).