Move all legacy WS protocol types into LegacyWsProtocol object
The entire JSON-RPC protocol and everything that supports it will
hopefully be removed soon after the first client release that uses the
new Protobuf protocol, though it will depend on how quickly users actually
update.

In the meantime, changing the namespacing like this just does that
little bit extra to ensure it's as clear as can be where legacy types
are still in use.