Enhance model to support multiple member owner public keys
As per https://developers.google.com/protocol-buffers/docs/encoding this
change can be made without requiring any migration because repeated and
non-repeated fields are encoded in the same way; the only difference is
in the generated code and read/write semantics.

For now BoardGame is updated to still require that a member has exactly
one public key (if it doesn't it simply won't appear in games). This
combined with the validation that a member has at least one public key
means that the legacy protocol can still be supported. Once clients
running the new protocol are released and following a deprecation period
the contraints may be relaxed and BoardGame may be enhanced.