Remove ZoneTerminatedNotication
This was an implementation detail that should never have been exposed
to clients. In the event that a ClientConnectionActor fails (most likely
because the node it was on is terminated), clients will obviously find
their connection is terminated. The most likely cause of a
ZoneValidatorActor termination is also that the node it is on is
terminated. In such cases the response of clients to that termination
is not actually the most appropriate action (they automatically attempt
to rejoin). This is because in such cases there will most likely be a
relatively significant delay before the shard region of the
ZoneValidatorActor in question is brought up again, and any join command
would be pending until it was.

Instead of leaking this implementation detail to clients what we now do
is bind the connection to the lifecycle of every zone that the client
joins, with the effect being that if a zone is terminated, so is the
connection (just as would be the case if the ClientConnectionActor's
node was terminated). So now clients will react in the same way -
reporting the failure to the user and requiring that they tap to attempt
a reconnect (future client versions may automate this, with an
increasing backoff and up to a maximum number of retries).