Eliminate use of AtLeastOnceDelivery
This primarily existed as an exercise in designing the zone-validator
protocol so that it accomodates the fact that messages between actors
can be lost, rather than simply appealing to the documentation which
points out that within one JVM delivery is effectively guaranteed (at
the time of implementing it only ever was deployed as one process; the
split to deploying separate containers for different roles came much
later).

Time and experience have revealed to me the downsides of using
AtLeastOnceDelivery like this (primarily the complexity it adds), and
I've come to realise that the better way to account for the fact that
messages may be dropped is to do something simpler: either push the
concern further out to the edges of the system, or recognise that
dropped messages are rare and that the consequences might not be so bad.

Hence this change, which takes both approaches:
1. For commands and responses there now simply is no retry of delivery.
   In practice this will be fine because if a command is dropped, users
   will simply retry once they conclude it hasn't been processed. If a
   response is dropped and the response was for a succesfully processed
   command, the consequence is that the client side RPC abstraction may
   leak a Promise (unless it has its own timeout mechanism). If the
   response was for a rejected command the additional consequence is
   that the user won't see any sort of error message - so we can expect
   they would retry, and would eventually get a duplicate of the same
   error.
2. For notifications there is no way for clients to know that they
   missed a single notification. But what we can do is detect gaps in
   notifications (and it is essential that we at least do that) to
   ensure that clients don't end up in inconsistent states. So, given
   that such cases should be very rare anyway, we use notification
   sequencne numbers and make the client-connection actor detect gaps
   and handle them by closing the connection; the user can then simply
   click to reconnect (the client can/may automatically retry in a
   future version).