Fix dead-client-removal regression
In the rare event that a zone validator is stopped while clients are
connected to it there are client joined events without matching client
quit events. On recovery this results in the set of connected clients
being non empty. Prior to the conversion to akka-typed this was handled
by triggering watches and unwatches for client connections regardless of
whether we were processing an event from a just-accepted command or
during recover, with the effect being that after recovery we would get a
Terminated message for each dead-client.

During the conversion to akka-typed these calls were moved out of
applyEvent because the current PersistentBehaviour implementation does
not permit applyEvent to be given a context to make watch/unwatch calls
against. The result of this was that any zones that were stopped while
clients were still connected would end up recovering with non-empty
connected client sets.

The fix is to workaround this limitation of PersistentBehaviour by
spawing a child actor to do the watching for us (we _are_ able to pass a
ref to another actor to applyEvent).