Fix zone validator passivation regression
The workaround for the lack of ability to wrap persistent behaviour
(spawning the peristent behaviour as a child of course) means that we
need to watch that child in order to ensure we passivate when it does.
Otherwise the sharded entity itself won't be passivated, and any
messages sent after the persistent entity has passivated will just go to
dead letters.