Switch to using eventsByTag for efficiency
Now that the analytics projection function doesn't require previous
state (it just reads what it needs when it needs it), the analytics
functionality as a whole can be made more efficient and the processing
delay reduced, by switching to eventsByTag.

The drawback of the previous approach in that regard was that it created
and kept open one stream for _each_ zone. The issue with that was that
because akka-persistence-cassandra ultimately just serves each stream by
polling Cassandra, it increased the load on Cassandra in proportion to
the total number of zones (even though at any given moment we expect
only a very small number to receive updates). The extent to which it
did this had led to the polling interval needing to be increased from 3s
to 300s in order to keep the average CPU usage below the rate at which
credits accumulate on a t2.small. This of course meant that there was a
delay of up to 5 minutes for activity to appear on the analytics side.

The new approach of using eventsByTag means we only keep _one_ stream
open, and so should reduce CPU usage such that the default polling
interval of 3s is sustainable again.

ZoneEventAdapter is added and configured to tag all future events, while
add-zone-event-tags.sh is added as a tool for "migrating" all existing
events (it works on a CSV export of the messages table by swapping
"null" in the 14th column for "zone-event" using gawk [gawk rather than
awk because the former supports in-place edits while the latter does
not]).