s/ActorRef/String/ in Client{Joined,Quit}Event
The problem with the former approach is that if/when journal migrations
are performed, the strings stored in the Protobuf messages all get
resolved and rewritten as dead-letters (due to the migration not even
being run in a cluster and so it being clear that the client-relay node
doesn't even exist; no network IO involved) - i.e. we end up
inadvertently rewriting history as part of the migration.