Set implementation module versions to _just_ the commit hash (no Git tag)
The semantic versioning in the tags is really only meant to describe the
protocol version - not the implementation. So for that reason alone the
implementation module versions shouldn't really include the tag (the
_implementation_ in the 2.0.0 commit itself for example wasn't actually
a breaking change at the point, because support for the 1.x.y protocol
was only dropped several weeks after the 2.0.0 client had been
released).

The additional benefits (which motivate this change _now_) are:
1. This also eliminates the need to replace + with - when using the
   version as a Docker image tag.
2. It means that the tag of the published Docker image can be determined
   with git describe alone (as opposed to having to call sbt version and
   parse the output, which would be excessively complicated because SBT
   doesn't provide a way to just print a single setting key to stdout).