Prefix Docker image tag with "evergreen-"
This is primarily done as a workaround for the fact that AWS ECR
Lifecycle Policies do not allow the set of tag prefixes in a tagged
image lifecycle rule to be empty (with the intended effect being to
apply the rule to all tagged images in the repo regardless of tag name),
and _nor_ do they allow the set of tag prefixes to exceed 10 characters
(to workaround the former limitation by listing each valid hex char,
i.e. 0,1,2,3,4,5,6,7,8,9,a,b,c,d,e,f which may appear at the start of a
Git SHA used as a Docker image tag).