Increase integration test patience config timeouts to 30s
(Just for cluster formation and service status).

Use akka-stream keepAlive for zone event-stream ping generation

PingGeneratorActor would also be deleted in this commit were it not for
the need to support the soon-to-be-legacy WS protocol* until most users
have updated to the 3.0.0+ client. So this is a simplification, but not
the full extent of it - that will come later (and when the existing TODO
re. testing is addressed, coverage of ping generation will be restored).

*which _could_ also be updated to use this approach, allowing
PingGeneratorActor to be removed _now_ (but won't be because there's no
point changing something that should soon be deleted).