Convert LiquidityServerSpec to cover v3 protocol
According to the Google Play Console, 66.4% of active users are now
running a v3.0.0 or higher client - so the new protocol is very much
_not_ experimental anymore; any regressions in it would matter far more
than any regressions in the v2 protocol. These tests could have been
updated in a way that would continue to cover the v2 protocol, but given
that it will soon be dropped entirely, it simply wouldn't have been
worth the extra time.