Remove deprecated websocket protocol
70% of active clients are now on version 3.0.0 or higher.

The logs show that there are people still using version 2.x.y clients,
but they've had more than long enough to update. When this change is
deployed they will get a connection error with this helpful hint:

"Ensure you are using the latest version of Liquidity" - so there's no
reason to wait any longer to do this cleanup.