Don't republish status messages
NotInfluenceReceiveTimeout as introduced in d46fb123 is not working as
it should, and I suspect it's because of
https://github.com/akka/akka/issues/24770.

Removing the periodic resending of status messages will eliminate the
issue entirely. It's OK to remove it because it was only added as a
precaution against message loss and the reality is that loss of these
messages is not of consequence - they are only for the /status/verbose
endpoint anyway, and in most cases fresher status messages are
constantly being sent anyway.