Move DNS record set from infrastructure stack to application stack
This will allow two infrastructure stacks for the same environment but
in different regions to be created without any resource conflicts. This
way an infrastructure stack can be prepared in a new region before
anything is removed from the old one, then the data can be migrated,
then the old application stack deleted (thus removing the records so
they can be created by a stack in the new region), and finally, the new
stack created.