Replace Gatling-based post-deploy tests with ScalaTest-based tests
It's unlikely that there's any future in which there is value in using
Gatling on Liquidity. The only reason for having initially used it as
the base for the post-deploy test was as a means of learning more about
Gatling; _not_ because it was an appropriate choice for Liquidity.

While I have not yet learned all that I ideally would know about it, I'm
not going to continue using Liquidity and my evenings and weekends as
the vehicle for doing so. By making this switch now the one outstanding
TODO that there is value in addressing (making the tests cover the event
stream too) is effectively solved for free (because the component tests
which are now used as a base for the post-deploy test already had a
mechanism for testing it).