Optimise analytics count maintenance
The previous triggers had intractable time complexity when reprojecting
the analytics tables (zone counts, member counts and account counts were
fine in practice, but public key counts took many minutes and the
transaction counts would have taken hours/days).

The reason the triggers were written to update every count that occured
after the inserted row was because there's a certain point in history
before which the order that events emerge from the journal does not
match the timestamps on the event envelopes - which results in the
counts being non-monotonic when naively updated via adding one to the
most recent count. Re-evaluating *every* count for a given entity type
after the time of the just-seen event was a workaround for this that
should have been tractable in practice based on the assumption that only
a minority of events occur out of total order. The hours+ running time
when the transaction trigger is included suggests that it is in fact the
majority of events that are only in partial order (ordered within the
scope of a zone, but lacking order across zones) - and it's this
excessive lack of order that makes the time complexity too high in
practice. Hence it's necessary to remove this workaround from the
triggers and make them assume the events are in total order (even though
they aren't), with the consequence being that the derived totals are
actually incorrect (my theory is that the lack of total order is a
result of a previous journal migration which was done one persistenct
entity at a time, as opposed to in total order without regard for the
persistent entity IDs - a difference resulting from the journal
previously having been stored in Cassandra). So either this
incorrectness will now just have to be accepted, or it may be resolved
by performing a journal migration which re-orders the events into total
order.

Also:
1. Simplify CFN helper script usage (remove need for create/update arg).
2. Add client session counts.