Combine redundantly isolated SBT modules
Back when the Scala model and Proto bindings were used by the Android
client this separation was necessary in order to expose only the
ws-protocol parts of the Scala model and Proto bindings. But since the
Android client now uses only the ws-protocol .proto files this isolation
has become redundant.

The definition of the Scala and Proto types still has to be done in a
separate module to that where the bindings are generated though (for the
reason described in 2b2fcc80). Hence keeping the Scala and Proto types
themselves in the model module, i.e. separated from the server module
(which is where the binding generation now happens).