Add devices.created analytics column, remove count tables
created should alway have been there - its omission was an error.

The count tables were always a questionable optimisation - the
duplication would make the counts inaccurate in the event that the main
projection tables were modified to flag test zones as not for inclusion
in analytics. Fortunately they're redundant now that the Grafana graphs
are showing counts-in-interval rather counts-to-date.