Make use of ScalaPB 0.8.0's new sealed oneof support
See https://scalapb.github.io/sealed-oneofs.html.

By making use of this the requirement for the wrappedOneofProtoBinding
derivation is almost eliminated. This simplifies the search for
ProtoBinding derivations and so should reduce compilation times, and
brings things one step closer to a state where the
wrappedOneofProtoBinding derivation could be removed entirely. That
would make the remaining ProtoBinding derivations simpler to reason
about, and perhaps also make the derivations simple enough that a
potentially faster alternative to Shapeless such as Magnolia might be
able to provide all required derivations.

The reason wrappedOneofProtoBinding has to remain (at least for now) is
because it's used for ZoneResponse ProtoBinding derivations - which is
necessary because ScalaPB doesn't support nested sealed oneofs (hence
the error vs. success cases of each ZoneResponse type still being
unsealed oneofs).