Make ZoneAnalyticsActor's close-stale-connections behaviour actually work
The initial cleaning action had been done manually, and since no stale
connections had manifested since that point, it wasn't apparent in the
Grafana dashboard that the automated cleaning function wasn't actually
doing anything. The MySQL 8.0 upgrade (and the re-projection to the
journal that entailed) made that become clear.

The reason it wasn't working was because the IO was never actually
executed; the unsafeToFuture/unsafeRunSync call was never made.

This could have been naively fixed with the one line addition of either
of the above, but the issue with unsafeToFuture would be that any errors
wouldn't get logged properly, and the issue with unsafeRunSync would be
that it would block the Akka dispatcher. Also, in the event of a
failure, there'd be no retry until the next scheduled run one hour
later.

Hence restructuring the close-stale-connections behaviour such that it's
implemented as a Source that gets combined with the existing
project-zone-events Source. This way errors are logged properly, the
Akka dispatcher is not blocked, and failures will result in retries with
an exponentially increasing backoff.

The removal of INTERVAL 1 DAY and INTERVAL 1 HOUR from the expression is
because now that the behaviour is actually running, it's apparent that,
unlike MySQL, H2 doesn't support the INTERVAL function - so
ZoneAnalyticsActorSpec was failing.

The removal of the NOT IN (active connection actor refs) clause is
because MySQL was rejecting the update in such cases due to a syntax
error, and given that this approach would be subject to a race condition
due to the removal of the mitigating intervals anyway, the simpler
solution is to simply remove this clause too and accept the possibility
of some connections being prematurely marked as closed in the analytics
table. That's OK, because if any are prematurely closed, the actual
close (and resulting quit event) will result in quit being updated to
the correct value.