Remove ZoneAnalyticsActor's close-stale-connections behaviour
As per the message in 55d332cf, there was a known race condition for
which there was no simple prevention, but which wasn't expected to be
much of a problem in practice. That expectation has been proved wrong;
in the time since that change was made I've had to truncate the
analytics tables several times due to not-actually-stale connections
getting closed and causing the ZoneAnalyticsActor to then fail when
attempting to project the later ClientQuitEvents.

The simplest solution is to revert back to doing
stale-connection-closing manually.