Setup ECS and ECR VPC endpoints, disable AssignPublicIp
It's always been desirable to avoid assigning each ECS task instance a
public IP, but it's only recently become possible with the release of
Fargate support for PrivateLink:
https://aws.amazon.com/about-aws/whats-new/2019/02/aws-fargate-now-has-support-for-aws-privatelink/

(The previous change was required before this could be done because each
VPC can only have one VPC endpoint for a given service, which meant
that a given AWS (account, region) would only be able to support one
environment - unless each environment was first made to have its own
VPC).

Also as part of this, on the same principle-of-least-privilege theme:
Restrict egress from task instances to only the required destinations
and ports.