Switch to in-JVM TLS termination
I was wrong when I stated in the commit message of 7468cb61 that ACM
allows imported certificates to be exported such that they could be used
for in-JVM TLs termination. It does allow the certificate to be
exported, but it doesn't allow the private key to be. Hence the need for
another place to store the cert and key; S3. So this commit does two
things:

1. Rewrite the Lambda to load and save the certbot config and data
   from/to S3 (this is also better in other ways - fewer LoC, deference to
   certbot itself when determining whether to renew, and retention of
   the LetsEncrypt account across runs).
2. Add an HTTPS binding to LiquidityServer and update the NLB to just
   work at the level of TCP.

Next steps:
1. Refine protocols and cipher suites.
2. Enable HTTP2.
3. Build signaling mechanism to trigger reload of certificates after
   renewals.