Retain symbolic links in certbot state
certbot was failing to read the previous certs with the following error:
certbot.errors.CertStorageError: expected
/tmp/config-dir/live/api.liquidityapp.com/cert.pem to be a symlink

It was then proceeding with the provisioning of new certificates - up
until the point that the Let's Encrypt rate limits were hit - only at
that point did failure become apparent (in the form of dead letters
being delivered as email via SNS).

Since certbot requires that the contents of the live directory all be
symlinks, and since S3 sync doesn't support commuting symlinks in and
out, another mechanism was needed... hence tar to the rescue.