Send Empty Zone notifications to GRPC subscribers to keep channel alive
The default GRPC settings
(https://github.com/grpc/grpc/blob/master/doc/keepalive.md) aren't
sufficient, and it's simpler and makes the intent clearer to do this at
the application level (plus, the akka-grpc docs don't mention how to
change the server side settings, and if there is an API, it isn't
easily discoverable).

Also increase idle timeout to 15 seconds in order to permit increased
keepalive interval of 10 seconds for both REST and GRPC notifications
(the compromise is that REST client disconnects will take longer to
detect, but since the REST API will hopefully be deprecated soon - at
which point the idle timeout can be restored to its default - that's
OK).